refactored testHelpers.js (previously functions.js), removed few linting errors
